useractions matches style guide
