+++ frontend/src/components/regions/UserActions/GithubActivity.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 
-export default ({ activityLog }) => {
+export default function GithubActivity({ activityLog }) {
   return (
     <Paper>
       <Grid container>
@@ -37,4 +37,4 @@ export default ({ activityLog }) => {
       </Grid>
     </Paper>
   );
-};
+}
